smartStoreB/requirements.txt
Django==6.0.2
mysqlclient==2.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1